Add missing labelClassName to column shape prop-types
Also defaults to `className`. This property is required when defining a column because it will serve as the entrypoint for defining column widths.

Given the fact that this tree-view uses flex-box as a means to layout the grid, it requires some user-defined widths. Perhaps I can work out some `flex-grow` shenanigans but I feel that will not work out as expected.