Enforce all children of <TabList> being a <Tab>
Because of how we manage focus for tabs, the children of the tablist cannot be a native DOM element, since React doesn't want us to apply any of our tab props to them. We could just skip applying props to native elements, but then our focus management stuff breaks here.

The simplest solution is to enforce that these are in fact tabs.