Configure whether the tab panel should be focusable or not
Resolves https://github.com/juanca/react-aria-components/issues/7.

Added after reading https://github.com/w3c/aria-practices/issues/323#issuecomment-400011586 and https://www.w3.org/TR/wai-aria-practices-1.1/#tabpanel.

In particular:

> When the tab list contains the focus, moves focus to the next element in the page tab sequence outside the tablist, wh
ich is typically either the first focusable element inside the tab panel or the tab panel itself.

Basically, if anything in the tab panel is focusable, then the tab panel doesn't need to be able to receive focus. However, if _*nothing*_ in the tab panel is focusable, the tab panel **does** need to receive focus. This is so that users using the keyboard for navigation won't inadvertently skip over the entire tab panel if the next focusable element is outside of the content.