Grid appropriately manages keyboard navigations
First, it should not default to "scrolling" the grid since the focus management should dictate the scrolling position.
Second, the management of indices was buggy: ensure the indices stay within the range of the axis -- no more, no less.
Last (but not least), use the default event for managing keyboard interactions (keydown).