Implement naive focusing mechanics on keyboard navigation
A few things wrong with this (but I just want to get this out there):

`ref` should *not* be used as a shortcut for focusing. It is clearly wrong since the `ref` is invoked with `null` reference. In addition, newer `ref` patterns use a value instead of a function.

While `focus()` provides an easy way of scrolling an element into "view" (and applies the appropriate focus state), it does not take into account the sticky elements. There is a case when the "focused" cell is hidden underneath a sticky cell.

It seems that this grid is slow when navigating on 20 * 200 cells -- in development mode, it is fine in production mode. There might be room for improvement by hoisting the focus management to the row and perhaps also a short-circuiting the rendering costs.