Switch focus management strategy: populate a list of refs that can be focused.
This replaces the previous strategy of passing down props that represented the coordinates of the focused gridcell. This was not performant (in my development environment).
On every key stroke, it would have to recalculate the whole React Element tree for the grid which is expensive at ~20 rows with ~200 columns.
I did try to short-circuit the rendering flows but I guess 2000 function calls is still a bit too much.

This new strategy is preferred since the grid is a tabular arrangement of components.
While this commit is pretty rough (WIP, end of day type of deal), it serves as the foundation for manipulating gridcells with focus and blur handlers.
Each gridcell can manage its own state thus its own tabindex.

I still have to figure out the "edit" mode of each cell but perhaps it is down the same path.