Try out the new React.forwardRef api
To be honest, I do not like this as much as the previous method - sending custom props down to child components with prop objects. Having said that, I am having some issues using this component in other projects do to React thinking that we are putting refs on stateless components. We are not, but here I just want to see if this eliminates the error.