Missing support for `relatedTarget` in IE11
It seems that this is only broken for IE9 -> 11. But supported pre-9. Wut.

Anyway, the tagged issue has some insight into workaround and a link to a longer thread about supporting a different kind of event (`focusin` and `focusout`).

I want to nail down some interactions in modern browsers first but I might have to revisit this. Either through graceful UX degradation or a different architecture for handling focus management -- might be more complicated and/or slower.