GridCell has an interactive mode
The GridCell uses a render prop for its children.
This allows a consumer to choose their own state when a cell should be in interactive mode.
I am not fully happy with the render prop but it seems to be the most flexible.
For example, the header cells do not have an interactive mode so there is currently a weird limbo state when those cells are in such mode -- due to the code.