Allow focus on grid container instead of a default focusable grid cell (on mount)
Because tab-index management is pushed down to grid cells (due to performance issues), it seems like it is impossible to have a grid cell be focusable by default on first mount. The "default" grid cell will never get the blur event which manages the tab-index of -1.

This is an okay trade off. We can even unset the tab-index of the grid once a cell receives focus too -- but let us iterate for the time being.