Bugfix: allow arrow keys to perform their default behavior
This allows selecting all the text in an input. I added the original line of code to prevent the table from scrolling with the arrow keys when the cell was in interactive mode AND the container element was the focused element. The real fix to the original problem was to focus the input.

Moving forward, any new interactive cell will need to ensure that an input (or control) element should be focused while in interactive mode.