Rename Grid to DataGrid
There are two types of grids: data grid and layout grid.

Data grid represents tabular interactive content -- headers and all.

Layout grid represents a composite widget for ensuring it only appears on the tab sequence once (just like a data grid) but it lacks any restrictions of being `tabular`. For example: the tabs list is a form of a layout grid.