Implement more in-depth interactive input cell
GridCell now re-renders when it is active -- not just when activating and deactivating.
GridCell can handle keyboard events -- and ignoring a false negative lint rule. Maybe this is an indicator that this is not a good design?

InputCell changes focus target when entering and exiting interactivity.

Interactive keyword kinda seems overloaded. Is there a better term?