feat(combobox): implement approximate suggestion feature
It only suggests once. In order to track against user typing then we would need an input that implement user-typing tracking. I am not really interested in recreating an input so I will skip it. Ideally, the input used in a combobox component is provided as a prop so a better combobox is composable.